Adds enum for QueueEntry.REQUIRES configuration
Previously custom queue entries had to specify "supported" shapes
in the `BaseQueueEntry.REQUIRES` class variable using list of names of
these shapes.

It makes it possible to specify these names with an enumeration instead
to make it more readable and better documented.

Also changes documentation to reflect on this change;
now in the code snippet describing creating custom queue entries, it
uses the new way of doing it.

It does not break old code; client application receives same string
values as before in the list