ISARAMaint: check if it's possible to read attributes
When setting up polling of the tango attributes, check that it
possible to read them. If there is issues, log it and run in a
'fault' mode.

This way the hardware object can handle situtation when ISARA's
tango device is unavailable or reports issues. For example if
the sample changer robot is turned off.

Without this checks, it was problematic to run MXCuBE, if the
robot was unavailable.