Tango Poller: handle numpy array shape changes
Add support for polling tango attribute values, which are numpy
arrays that change shapes.

Currently if, for example, a tango attribute changes value
from [1, 2] to [], poller will fail.

It will fail, because numpy does not support '==' operation on
arrays of different 'shapes'.

Adds code to explicitly check if shapes are different between old
and new values.