Remove unecessary function call, add signal emiting at set_limits()
In the parent class, AbstractNState, the functions set_limits and
update_limits raise a NotImplementedError. Because LNLSZoom is a
class that is an N state but has limits, the limits must be set
manually. This is why the limits are defined at init (and not at
the configuration file) and also set with specific functions at
the LNLSZoom class. There could be a way to set these limits in the
configuration file, but I could not find that.